Allow all new types to be cached and loaded.
All new types are tagged with a PythonCachedTypeAttribute, so that when
they are saved to disk they can be loaded using
NewTypeMaker.LoadCachedTypes.

Previously, only types stored with NewTypeMaker.SaveTypes (or
clr.CompileSubclassTypes) were loaded from assemblies.

The main advantage to this is that types will not need to be generated at
runtime if they are alreadfy availble.